Partial Health System
Added a health script. It currently defaults to 100 health and displays
curhealth/maxhealth. The #'s can be changed later, as well as the health
bar looks and position itself. Right now it's just on the top. Also the
bar changes size based on the current health remaining. Need to still
set a timer and reset the invulnerability based on the timer.