Big update: Dialogue scripts are now a full cutscene system.
Add actions to the CutsceneManager class to allow you to trigger any
in-game event from the script. You can also pause a script and
wait for events from the game to proceed. This should allow me to
write all cutscenes in these separate cutscene script files and simply
play them in the game at the appropriate times.